Changing the save method to be a external callable function in the menu
